LayoutTests/imported/w3c/web-platform-tests/css/css-values/tree-counting/sibling-function-invalidation-expected.txt


PASS Initially 6th sibling
FAIL 5th sibling after removal assert_equals: expected 50 but got 60
PASS Initially 6 siblings
FAIL 5 siblings after removal assert_equals: expected 50 but got 60
PASS Initially 6th slotted sibling
FAIL 5th sibling after slot change assert_equals: expected 50 but got 60
PASS Initially 6 slotted siblings
FAIL 5 siblings after slot change assert_equals: expected 50 but got 60
